==> experiment1/analysis_scripts/3_voc_RL.Rmd <==
---
title: "VoC Analyses Part 3: Analyze Reinforcement-Learning Results"
date: 3/27/24
output:
    html_document:
        df_print: 'paged'
        toc: true
        toc_float:
            collapsed: false
            smooth_scroll: true
        number_sections: false
        code_download: true
        self_contained: true
---

```{r chunk settings, include = FALSE}
# set chunk settings
knitr::opts_chunk$set(echo = FALSE, 
                      cache = TRUE,
                      message = FALSE,
                      warning = FALSE)
knitr::opts_chunk$set(dpi=600)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

```{r load packages, include = F}

# list all packages required for the analysis
list.of.packages <- c("tidyverse", "latex2exp", "afex", "sjPlot")

# check if all packages are installed, if not, install them.
new.packages <- list.of.packages[!(list.of.packages %in% installed.packages()[,"Package"])]
if(length(new.packages)) install.packages(new.packages)

# load all packages 
lapply(list.of.packages, library, character.only = TRUE)

# add theme for plotting
voc_theme <- function () {
  theme(
    panel.border = element_rect(fill = "transparent", color="gray75"),
    panel.background  = element_blank(),
    plot.background = element_blank(), 
    legend.background = element_rect(fill="transparent", colour=NA),
    legend.key = element_rect(fill="transparent", colour=NA),
    line = element_blank(),
    axis.ticks = element_line(color="gray75"),
    text=element_text(family="Avenir"),
    axis.text = element_text(size = 12),
    axis.title = element_text(size = 15),
    title = element_text(size = 15),
    strip.background = element_blank(),
    strip.text = element_text(size=12)
  )
}

color8 = "#80dbb2"
color1 = "#00b4d8"
color2 = "#0077b6"
color3 = "#03045e"
color4 = "#84347C"
color5 = "#B40424"
color6 = "#EB6D1E"
color7 = "#f5b68f"

scale_this <- function(x){
  (x - mean(x, na.rm=TRUE)) / sd(x, na.rm=TRUE)
}

```

```{r, load data}
#load data
aics = read_csv("RL_modeling/output/aics_all_16_models_100iter.csv")
bics = read_csv("RL_modeling/output/bics_all_16_models_100iter.csv")
```

```{r pivot data longer}
aics1 <- pivot_longer(aics, 
                cols = oneAlpha_oneBeta:fourAlpha_twoBeta_agencyBonus,
                names_to = "model",
                values_to = "AIC")

bics1 <- pivot_longer(bics, 
                cols = oneAlpha_oneBeta:fourAlpha_twoBeta_agencyBonus,
                names_to = "model",
                values_to = "BIC")
```


#  AIC analyses
## Average AIC by age group
```{r plot AIC by age group, fig.width = 8, fig.height = 5, units = "in"}

# Add id and other demographic info
sub_info <- read_csv('data/voc_sub_info.csv') %>%
    mutate(age_group = case_when(age < 13 ~ "Children",
                                 age > 12.99 & age < 18 ~ "Adolescents",
                                 age > 17.99 ~ "Adults"))

sub_info$age_group <- factor(sub_info$age_group, levels = c("Children", "Adolescents", "Adults"))

model_results <- full_join(sub_info, aics1, by = c("subID"))

model_results$model <- factor(model_results$model, 
                              levels = c("oneAlpha_oneBeta",
                                         "oneAlpha_twoBeta",
                                         "twoAlpha_oneBeta",
                                         "twoAlpha_twoBeta",
                                         "twoAlphaValenced_oneBeta",
                                         "twoAlphaValenced_twoBeta",
                                         "fourAlpha_oneBeta",
                                         "fourAlpha_twoBeta",
                                         "oneAlpha_oneBeta_agencyBonus",
                                         "oneAlpha_twoBeta_agencyBonus",
                                         "twoAlpha_oneBeta_agencyBonus",
                                         "twoAlpha_twoBeta_agencyBonus",
                                         "twoAlphaValenced_oneBeta_agencyBonus",
                                         "twoAlphaValenced_twoBeta_agencyBonus",
                                         "fourAlpha_oneBeta_agencyBonus",
                                         "fourAlpha_twoBeta_agencyBonus"))
model_results <- model_results %>%
    mutate(agencyBonus = case_when(str_detect(model, "agency") ~ "With Agency Bonus",
                                  !str_detect(model, "agency") ~ "No Agency Bonus"),
           shortName = str_remove(model, '_agencyBonus'))

model_results$shortName <- factor(model_results$shortName,
                                  levels = c("oneAlpha_oneBeta",
                                         "oneAlpha_twoBeta",
                                         "twoAlpha_oneBeta",
                                         "twoAlpha_twoBeta",
                                         "twoAlphaValenced_oneBeta",
                                         "twoAlphaValenced_twoBeta",
                                         "fourAlpha_oneBeta",
                                         "fourAlpha_twoBeta"))
                                 
#summarize
model_summary <- model_results %>%
    group_by(age_group, shortName, agencyBonus) %>%
    summarize(meanAIC = mean(AIC))

# # Plot the results by age group 
AIC_age_plot <- ggplot(model_summary, aes(x = age_group, y = meanAIC, fill = shortName))+
    facet_wrap(~agencyBonus) +
    geom_bar(stat = "identity", position = "dodge", color = "black") +
    scale_fill_manual(name = "Model",
                      values = c(color8, color1, color2, color3, color4, color5, color6, color7, color1),
                      labels =  c(TeX('$one\\alpha\\_one\\beta'),
                                TeX('$one\\alpha\\_two\\beta'),
                                TeX('$twoChoice\\alpha\\_one\\beta'),
                                TeX('$twoChoice\\alpha\\_two\\beta'),
                                TeX('$twoValenced\\alpha\\_one\\beta'),
                                TeX('$twoValenced\\alpha\\_two\\beta'),
                                TeX('$four\\alpha\\_one\\beta'),
                                TeX('$four\\alpha\\_two\\beta'))) + 
    coord_cartesian(ylim = c(350, 600)) +
    ylab("Mean AIC") +
    xlab("") +
    voc_theme() +
    theme(axis.text.x = element_text(angle = 60, hjust = 1))
AIC_age_plot
```

## Average AIC 
```{r aic overall plot, fig.width = 6, fig.height = 4, units = "in"}
model_summary_overall <- model_results %>%
    group_by(model, shortName, agencyBonus) %>%
    summarize(meanAIC = mean(AIC))

AIC_plot <- ggplot(model_summary_overall, aes(x = shortName, y = meanAIC, fill = shortName)) +
    geom_bar(stat = "identity", position = "dodge", color = "black") +
    facet_wrap(~agencyBonus) +
    coord_cartesian(ylim = c(350, 600)) + 
    ylab("Mean AIC") +
    xlab("Model") +
    scale_fill_manual(name = "Model",
                      values = c(color8, color1, color2, color3, color4, color5, color6, color7, color1),
                      labels =  c(TeX('$one\\alpha\\_one\\beta'),
                                TeX('$one\\alpha\\_two\\beta'),
                                TeX('$twoChoice\\alpha\\_one\\beta'),
                                TeX('$twoChoice\\alpha\\_two\\beta'),
                                TeX('$twoValenced\\alpha\\_one\\beta'),
                                TeX('$twoValenced\\alpha\\_two\\beta'),
                                TeX('$four\\alpha\\_one\\beta'),
                                TeX('$four\\alpha\\_two\\beta'))) + 
    scale_x_discrete(labels =  c(TeX('$one\\alpha\\_one\\beta'),
                                TeX('$one\\alpha\\_two\\beta'),
                                TeX('$twoChoice\\alpha\\_one\\beta'),
                                TeX('$twoChoice\\alpha\\_two\\beta'),
                                TeX('$twoValenced\\alpha\\_one\\beta'),
                                TeX('$twoValenced\\alpha\\_two\\beta'),
                                TeX('$four\\alpha\\_one\\beta'),
                                TeX('$four\\alpha\\_two\\beta'))) + 
    voc_theme() +
        theme(axis.text.x = element_text(angle = 75, hjust = 1),
              legend.position = "none")
AIC_plot

```

## AIC difference from best model
```{r aic overall difference plot, fig.width = 4, fig.height = 5, units = "in"}
#get minimum AIC
minAIC = min(model_summary_overall$meanAIC)

#subtract from mean AICs
model_difference_summary <- model_summary_overall %>%
    mutate(AIC_difference = meanAIC - minAIC[1]) %>%
    filter(agencyBonus == "With Agency Bonus")

#plot
AIC_difference_plot <- ggplot(model_difference_summary, aes(x = shortName, y = AIC_difference, fill = shortName)) +
    geom_bar(stat = "identity", position = "dodge", color = "black") +
    facet_wrap(~agencyBonus) +
    ylab("AIC Difference") +
    xlab("") +
    scale_fill_manual(name = "Model",
                      values = c(color8, color1, color2, color3, color4, color5, color6, color7, color1),
                      labels =  c(TeX('$one\\alpha\\_one\\beta'),
                                TeX('$one\\alpha\\_two\\beta'),
                                TeX('$twoChoice\\alpha\\_one\\beta'),
                                TeX('$twoChoice\\alpha\\_two\\beta'),
                                TeX('$twoValenced\\alpha\\_one\\beta'),
                                TeX('$twoValenced\\alpha\\_two\\beta'),
                                TeX('$four\\alpha\\_one\\beta'),
                                TeX('$four\\alpha\\_two\\beta'))) + 
    scale_x_discrete(labels =  c(TeX('$one\\alpha\\_one\\beta'),
                                TeX('$one\\alpha\\_two\\beta'),
                                TeX('$twoChoice\\alpha\\_one\\beta'),
                                TeX('$twoChoice\\alpha\\_two\\beta'),
                                TeX('$twoValenced\\alpha\\_one\\beta'),
                                TeX('$twoValenced\\alpha\\_two\\beta'),
                                TeX('$four\\alpha\\_one\\beta'),
                                TeX('$four\\alpha\\_two\\beta'))) + 
    voc_theme() +
        theme(axis.text.x = element_text(angle = 60, hjust = 1),
              legend.position = "none")
AIC_difference_plot

```


#  Age-related change in parameter estimates from models
```{r parameter estimates}

# load all parameters from each model
model_params <- read_csv("RL_modeling/output/model_fits_real_data/fourAlpha_twoBeta_agencyBonus.csv",
                         col_names = c("negLL",
                                       "logPost",
                                       "AIC",
                                       "BIC",
                                       "alphaPosChoice",
                                       "alphaNegChoice",
                                       "alphaPosComp",
                                       "alphaNegComp",
                                       "betaAgency",
                                       "betaMachine",
                                       "agencyBonus"))

#add sub ID and information
subID <- read_csv('RL_modeling/output/subIDs.csv')
model_params <- bind_cols(subID, model_params)
model_params <- full_join(sub_info, model_params, by = c("subID"))
```


# Run regressions between model parameters and age
```{r param age regressions}

model_params$LL <- model_params$negLL * -1

# Log likelihood
summary(lm(LL ~ age, data = model_params))
# significant

# Alpha Pos Choice
summary(lm(alphaPosChoice ~ age, data = model_params))
#not significant

# Alpha Neg Choice
summary(lm(alphaNegChoice ~ age, data = model_params))
#not significant

# Alpha Pos Comp
summary(lm(alphaPosComp ~ age, data = model_params))
#not significant

# Alpha Neg Comp
summary(lm(alphaNegComp ~ age, data = model_params))
#not significant

# Beta Agency
summary(lm(betaAgency ~ age, data = model_params))
#significant

# Beta Bandit
summary(lm(betaMachine ~ age, data = model_params))
#not significant

# agency bonus
summary(lm(agencyBonus ~ age, data = model_params))
#not significant
```

# Plot relations between model parameters and age
```{r age parameter plot, fig.width = 7, fig.height = 4, units = "in"}

params_long <- model_params %>%
    pivot_longer(names_to = "param",
                 values_to = "estimate",
                 cols = c(alphaPosChoice:agencyBonus)) 

params_long$param <- factor(params_long$param, 
                            levels = c("alphaPosChoice",
                                       "alphaNegChoice",
                                       "alphaPosComp",
                                       "alphaNegComp",
                                       "betaAgency",
                                       "betaMachine",
                                       "agencyBonus"),
                            labels = c(TeX("$\\alpha_{choice_+}$"), 
                                       TeX("$\\alpha_{choice_-}$"), 
                                       TeX("$\\alpha_{comp_+}$"), 
                                       TeX("$\\alpha_{comp_-}$"), 
                                       TeX("$\\beta_{agency}$"), 
                                       TeX("$\\beta_{machine}$"),
                                       "Agency~Bonus"
                                ))

params_plot <- ggplot(params_long, aes(x = age, y = estimate, color = param)) +
    facet_wrap(~param, scale = "free", labeller = label_parsed, nrow = 2) +
    geom_point() +
    geom_smooth(method = "lm", aes(fill = param)) +
    ylab("Parameter Estimate") +
    xlab("Age") +
    voc_theme() +
    theme(legend.position = "none")
params_plot
```

# Parameter summary statistics
```{r parameter summary stats}

param_summary <- params_long %>%
    group_by(param) %>%
    summarize(meanEstimate = mean(estimate),
            seEstimate = sd(estimate)/sqrt(n()))
param_summary

```

# Mixed-effects beta analysis
```{r beta regression}
betas <- model_params %>%
    pivot_longer(cols = c(betaAgency, betaMachine),
                 names_to = "betaType",
                 values_to = "estimate") %>%
    select(subID, age, age_group, betaType, estimate) %>%
    unique() 
                               
betas$ageZ <- scale_this(betas$age)

beta_model <- mixed(estimate ~ ageZ * betaType + (1|subID),
                             data = betas,
                             method = "S")
beta_model
summary(beta_model)
```

```{r beta print model stats}

beta_model.lmer <- mixed(estimate ~ ageZ * betaType + (1|subID),
                             data = betas,
                             method = "S",
                             return = "merMod")

tab_model(beta_model.lmer, 
          pred.labels = c("intercept", "age", "decision stage", "age x decision stage"),
          transform = NULL,
          show.est = T, 
          show.se = T, 
          show.stat = T,
          show.ci = F,
          show.re.var = F,
          show.icc = F,
          show.ngroups = F,
          show.obs = F,
          show.r2 = F,
          string.se = "SE",
          emph.p = F,
          string.pred = "Predictor",
          title = "",
          dv.labels = "")
```


## Beta plot
```{r beta plot}

beta_means <- betas %>%
    group_by(age_group, betaType) %>%
    summarize(meanBeta = mean(estimate),
              seBeta = sd(estimate) / sqrt(n()))

beta_plot <- ggplot(beta_means, aes(x = betaType, y = meanBeta, fill = age_group)) +
    geom_bar(color = 'black', stat = "identity", position = "dodge") + 
    geom_errorbar(color = "black", aes(ymin = meanBeta - seBeta, ymax = meanBeta + seBeta), width = .1,
                  position = position_dodge(width = .9)) +
    scale_fill_manual(values = c(color1, color2, color3), name = "Age Group") +
    ylab("Mean Beta") +
    xlab("Decision Stage") +
    scale_x_discrete(labels = c("Agency Decision", "Machine Decision")) +
    voc_theme()
beta_plot 


beta_plot_continuous <- ggplot(betas, aes(color = betaType, y = estimate, x = age)) +
    geom_point() +
    geom_smooth(method = "lm", aes(fill = betaType, color = betaType)) +
    scale_color_manual(values = c(color1, color2), name = "Beta Parameter", labels = c("Agency Decision", "Machine Decision")) +
    scale_fill_manual(values = c(color1, color2), name = "Beta Parameter", labels = c("Agency Decision", "Machine Decision")) +
    ylab("Beta Estimate") +
    xlab("Age") +
    voc_theme()
beta_plot_continuous
```


# Mixed-effects learning rate analysis
```{r learning rate regression}
learning_rates <- model_params %>%
    pivot_longer(cols = c(alphaPosChoice:alphaNegComp),
                 names_to = "learningRate",
                 values_to = "estimate") %>%
    select(subID, age, age_group, learningRate, estimate) %>%
    unique() %>%
    mutate(valence = case_when(str_detect(learningRate, "Pos") ~ "Positive",
                               str_detect(learningRate, "Neg") ~ "Negative"),
           agency = case_when(str_detect(learningRate, "Choice") ~ "Choice",
                              str_detect(learningRate, "Comp") ~ "Comp"))
                               
learning_rates$ageZ <- scale_this(learning_rates$age)

learning_rate_model <- mixed(estimate ~ ageZ * valence * agency + (1|subID),
                             data = learning_rates,
                             method = "S")
learning_rate_model
summary(learning_rate_model)
# valence x agency interaction
# marginal valence x agency x age interaction

#t test between alpha pos choice and alpha neg choice
t.test(model_params$alphaPosChoice, model_params$alphaNegChoice, paired = T)
#significant

#t test between alpha pos comp and alpha neg comp
t.test(model_params$alphaPosComp, model_params$alphaNegComp, paired = T)
#not significant

```

```{r learning rate print model stats}

learning_rate_model.lmer <- mixed(estimate ~ ageZ * valence * agency + (1|subID),
                             data = learning_rates,
                             method = "S",
                             return = "merMod")

tab_model(learning_rate_model.lmer, 
          pred.labels = c("intercept", "age", "valence", "agency", "age x valence", "age x agency", "valence x agency", "age x valence x agency"),
          transform = NULL,
          show.est = T, 
          show.se = T, 
          show.stat = T,
          show.ci = F,
          show.re.var = F,
          show.icc = F,
          show.ngroups = F,
          show.obs = F,
          show.r2 = F,
          string.se = "SE",
          emph.p = F,
          string.pred = "Predictor",
          title = "",
          dv.labels = "")
```

## Learning rate plot
```{r learning rate plot}

learning_rate_means <- learning_rates %>%
    group_by(agency, valence) %>%
    summarize(meanLR = mean(estimate),
              seLR = sd(estimate) / sqrt(n()))

learning_rate_plot <- ggplot(learning_rate_means, aes(x = agency, y = meanLR, fill = valence)) +
    geom_bar(color = 'black', stat = "identity", position = "dodge") + 
    geom_errorbar(color = "black", aes(ymin = meanLR - seLR, ymax = meanLR + seLR), width = .1,
                  position = position_dodge(width = .9)) +
    scale_fill_manual(values = c(color1, color2), name = "Valence") +
    ylab("Mean Learning Rate") +
    xlab("Agency") +
    scale_x_discrete(labels = c("Participant Choice", "Computer Choice")) +
    voc_theme()
learning_rate_plot 
```



# Relation between parameter estimates and 'model-free' regressions
```{r relations between random effects and model parameters - extract REs}

# Read in data
banditTask <- read_csv('data/processed/bandit_task.csv') 

#combine with participant age
banditTask <- full_join(banditTask, sub_info, by = c("subID"))

#scale voc
banditTask$zVoC <- scale_this(banditTask$voc)
banditTask$zTrialOfCond <- scale_this(banditTask$trialOfCond)
banditTask$zAge <- scale_this(banditTask$age)

# predict agency choice from utility of control, trial, linear age
agency_byVOCTrialAge.mixed = mixed(agency ~ zVoC * zTrialOfCond + (zVoC * zTrialOfCond|subID), 
                        data = banditTask, 
                        family = binomial, 
                        method = "LRT", control=glmerControl(optimizer="bobyqa",optCtrl=list(maxfun=1e6)),
                        return = "merMod") 

#get random effects
voc_REs <- ranef(agency_byVOCTrialAge.mixed)$subID %>%
    rownames_to_column(var = "subID")

#combine with RL estimates
voc_REs_RL <- full_join(voc_REs, model_params, by = 'subID')

```

```{r run regressions REs and model parameters}
#run regressions

#agency bonus
voc_intercept_agencyBonus.lm <- lm(`(Intercept)` ~ agencyBonus, data = voc_REs_RL)
summary(voc_intercept_agencyBonus.lm)

#beta agency
voc_slope_betaAgency.lm <- lm(zVoC ~ betaAgency, data = voc_REs_RL)
summary(voc_slope_betaAgency.lm)

#beta agency controlling for age
voc_slope_betaAgencyAge.lm <- lm(zVoC ~ betaAgency + age, data = voc_REs_RL)
summary(voc_slope_betaAgencyAge.lm)

#beta agency controlling for beta machine
voc_slope_betaMachine.lm <- lm(zVoC ~ betaAgency + betaMachine, data = voc_REs_RL)
summary(voc_slope_betaMachine.lm)

```







# Questionnaire relations

## DOC
```{r doc}
# load questionnaire data
DOC <- read_csv("data/scored_surveys/DOC_scored.csv", col_names = TRUE) 

# merge with model params
DOC <- left_join(DOC, model_params)

# z score continuous variables
DOC$zAge <- scale_this(DOC$age)
DOC$zBetaAgency <- scale_this(DOC$betaAgency)
DOC$zAgencyBonus <- scale_this(DOC$agencyBonus)

# relation between DOC and age
lm(DOC ~ zAge, DOC) %>% summary()
#marginal positive effect (p = .058)

# relation between DOC and VoC
lm(DOC ~ zBetaAgency * zAgencyBonus *zAge, DOC) %>% summary()
# no effects

```

## LOC
```{r loc}
# load questionnaire data
LOC <- read_csv("data/scored_surveys/LOC_scored.csv", col_names = TRUE) 

# merge with model params
LOC <- left_join(LOC, model_params)

#z score continuous variables
LOC$zAge <- scale_this(DOC$age)
LOC$zBetaAgency <- scale_this(LOC$betaAgency)
LOC$zAgencyBonus <- scale_this(LOC$agencyBonus)

# relation between LOC and age
lm(LOC ~ zAge, LOC) %>% summary()
# no effect

# relation between LOC and VoC
lm(LOC ~ zBetaAgency * zAgencyBonus * zAge, LOC) %>% summary()
# no effects
```


## BDI
```{r bdi}
# load questionnaire data
BDI <- read_csv("data/scored_surveys/BDI_scored.csv", col_names = TRUE) 

# merge with model params
BDI <- left_join(BDI, model_params)

#z score continuous variables
BDI$zAge <- scale_this(BDI$age)
BDI$zBetaAgency <- scale_this(BDI$betaAgency)
BDI$zAgencyBonus <- scale_this(BDI$agencyBonus)

# relation between BDI and age
lm(zBDI ~ zAge, BDI) %>% summary()
# no effect

# relation between BDI and VoC 
lm(zBDI ~ zBetaAgency * zAgencyBonus *zAge, BDI) %>% summary()
# no effects

```


## STAI
```{r stai}
# load questionnaire data
STAI <- read_csv("data/scored_surveys/STAI_scored.csv", col_names = TRUE) 

# merge with model params
STAI <- left_join(STAI, model_params)

#z score continuous variables
STAI$zAge <- scale_this(STAI$age)
STAI$zBetaAgency <- scale_this(STAI$betaAgency)
STAI$zAgencyBonus <- scale_this(STAI$agencyBonus)

# relation between STAI_t and age
lm(zSTAI_t ~ zAge, STAI) %>% summary()
# no effect

# relation between STAI_t and VoC
lm(zSTAI_t  ~ zBetaAgency * zAgencyBonus *zAge, STAI) %>% summary()
# no effect

# relation between STAI_s and age
lm(zSTAI_s ~ zAge, STAI) %>% summary()
# no effects

# relation between STAI_s and VoC
lm(zSTAI_s  ~ zBetaAgency * zAgencyBonus *zAge, STAI) %>% summary()
# no effects
```
