==> experiment1/analysis_scripts/5_voc_posterior_predictive.Rmd <==
---
title: "VoC Analyses Part 5: Posterior Predictive Checks"
date: 3/27/24
output:
    html_document:
        df_print: 'paged'
        toc: true
        toc_float:
            collapsed: false
            smooth_scroll: true
        number_sections: false
        code_download: true
        self_contained: true
---
```{r chunk settings, include = FALSE}
# set chunk settings
knitr::opts_chunk$set(echo = FALSE, 
                      cache = TRUE,
                      message = FALSE,
                      warning = FALSE)
knitr::opts_chunk$set(dpi=600)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```
                      
```{r setup, include=FALSE}
# list all packages required for the analysis
list.of.packages <- c("tidyverse", "afex")

# check if all packages are installed, if not, install them.
new.packages <- list.of.packages[!(list.of.packages %in% installed.packages()[,"Package"])]
if(length(new.packages)) install.packages(new.packages)

# load all packages 
lapply(list.of.packages, library, character.only = TRUE)

# add theme for plotting
voc_theme <- function () {
  theme(
    panel.border = element_rect(fill = "transparent", color="gray75"),
    panel.background  = element_blank(),
    plot.background = element_blank(), 
    legend.background = element_rect(fill="transparent", colour=NA),
    legend.key = element_rect(fill="transparent", colour=NA),
    line = element_blank(),
    axis.ticks = element_line(color="gray75"),
    text=element_text(family="Avenir"),
    axis.text = element_text(size = 12),
    axis.title = element_text(size = 15),
    title = element_text(size = 15),
    strip.background = element_blank(),
    strip.text = element_text(size=12)
  )
}

```

```{r load data}
sim_data <- read_csv("RL_modeling/output/posterior_predictive/simData_subParams_fourAlpha_twoBeta_agencyBonus.csv") %>%
    rename(tokenOffer = offer,
           agencyResp = agencyChoice,
           banditResp = banditChoice,
           reward = outcome) %>%
    mutate(agency = case_when(agencyResp == 2 ~ 1,
                              agencyResp == 1 ~ 0),
           correct = case_when(leftBandit == "bandit50a" & banditResp == 1 ~ 1,
                               leftBandit == "bandit50a" & banditResp == 2 ~ 0,
                               leftBandit == "bandit50b" & banditResp == 1 ~ 0,
                               leftBandit == "bandit50b" & banditResp == 2 ~ 1,
                               leftBandit == "bandit30" & banditResp == 1 ~ 0,
                               leftBandit == "bandit30" & banditResp == 2 ~ 1,
                               leftBandit == "bandit70" & banditResp == 1 ~ 1,
                               leftBandit == "bandit70" & banditResp == 2 ~ 0,
                               leftBandit == "bandit10" & banditResp == 1 ~ 0,
                               leftBandit == "bandit10" & banditResp == 2 ~ 1,
                               leftBandit == "bandit90" & banditResp == 1 ~ 1,
                               leftBandit == "bandit90" & banditResp == 2 ~ 0),
           condition = case_when(leftBandit == "bandit30" ~ "bandits7030",
                                 leftBandit == "bandit70" ~ "bandits7030",
                                 leftBandit == "bandit90" ~ "bandits9010",
                                 leftBandit == "bandit10" ~ "bandits9010",
                                 leftBandit == "bandit50a" ~ "bandits5050",
                                 leftBandit == "bandit50b" ~ "bandits5050"),
           voc = case_when(condition=="bandits9010" ~ 9 - (tokenOffer + 5),
                           condition=="bandits7030" ~ 7 - (tokenOffer + 5),
                           condition=="bandits5050" ~ 5 - (tokenOffer + 5))) %>%
    group_by(subID, simID) %>%
    mutate(trial = row_number()) %>%
    ungroup() %>%
    group_by(subID, simID, condition) %>%
    mutate(trialOfCond = row_number()) %>%
    ungroup() %>%
    mutate(dataType = "Model Simulation")
    
sim_data$subID <- str_sub(sim_data$subID, end = -5)
```

```{r load real data}

#load demographic info
sub_info <- read_csv('data/voc_sub_info.csv') %>%
    mutate(age_group = case_when(age < 13 ~ "Children",
                                 age > 12.99 & age < 18 ~ "Adolescents",
                                 age > 17.99 ~ "Adults"))

sub_info$age_group <- factor(sub_info$age_group, levels = c("Children", "Adolescents", "Adults"))

#merge demographics with simulated data
combined_data <- full_join(sub_info, sim_data, by = c("subID"))

#load bandit data
banditTask <- read_csv('data/processed/bandit_task.csv') %>%
     mutate(dataType = "Empirical",
           simID = 0)

#merge demographics with readl data
banditTask <- full_join(sub_info, banditTask, by = c("subID"))


#merge real data
empirical_data <- banditTask %>% 
    select(subID, gender, age, age, age_group, simID, leftBandit, rightBandit,
           tokenOffer, agencyResp, banditResp, reward, agency, correct = stage_2_acc,
           condition, voc, trial, trialOfCond, dataType)

all_data <- rbind(combined_data, empirical_data) %>%
    mutate(block = floor((trial-1)/21 + 1))
```


## Plot: Proportion optimal machine selections across age groups and trials
```{r plot bandit choices across trials, width = 7, height = 4, unit = "in"}

#plot simulated data
all_data %>% 
  filter(agency==1, condition!="bandits5050") %>% 
  group_by(condition, block, age_group, dataType) %>% 
  summarize(pctCorr = mean(correct), 
    se = sd(correct)/sqrt(n())) %>% 
  ggplot(., aes(x=block, y=pctCorr, color=condition)) +
  geom_point() +
    geom_smooth(method = "lm", aes(fill = condition)) +
    geom_hline(yintercept = .5, linetype="dashed") +
    ylab("Proportion Optimal Machine Selections") +
    xlab("Block") +
    facet_grid(cols = vars(age_group), rows = vars(dataType)) +
    scale_fill_manual(name="Condition",
                      labels=c("70/30",
                               "90/10"),
                      values=c("#0077b6", "#03045e"), 
                      guide = guide_legend(reverse=TRUE)) +
    scale_color_manual(name="Condition",
                      labels=c("70/30",
                               "90/10"),
                      values=c("#0077b6", "#03045e"),
                     guide = guide_legend(reverse=TRUE)) +
    voc_theme() +
    theme(strip.text = element_text(size=12))
```


## Compute average error in agency predictions
```{r posterior predictive error in machine selections}

sim_machine_means <- all_data %>%
    filter(agency==1, condition!="bandits5050") %>% 
    filter(dataType == "Model Simulation") %>%
    group_by(subID, age, age_group, block, condition) %>%
    summarize(simMeanCorrect= mean(correct, na.rm = T))

VoC_sub_error_machines <- all_data %>%
    filter(agency==1, condition!="bandits5050") %>% 
    filter(dataType == "Empirical") %>%
    select(subID, age, age_group, block, condition, correct) %>%
    left_join(sim_machine_means, by = c("subID", "age", "age_group", "block", "condition")) %>%
    mutate(error = simMeanCorrect - correct,
           absError = abs(error))

VoC_mean_error_machines <- VoC_sub_error_machines %>%
    group_by(age_group, block, condition) %>%
    summarize(meanError = mean(error, na.rm = T),
              seError= sd(error / sqrt(n())),
              meanAbsError = mean(absError, na.tm = T)
              )
```

## Plot: Error in  optimal machine selections across age groups and trials
```{r plot model error bandit choices across trials, width = 7, height = 4, unit = "in"}

#plot simulated data
ggplot(VoC_mean_error_machines, aes(x=block, y=meanError, color=condition)) +
  geom_point() +
    geom_line() +
  geom_hline(yintercept = 0, linetype="dashed") +
  geom_errorbar(aes(ymin = meanError - seError, ymax = meanError + seError), width = .1) + 
  voc_theme() +
  ylab("Mean Error in Model Predictions") +
  xlab("Block") +
  facet_grid(cols = vars(age_group)) +
      scale_fill_manual(name="Condition",
                      labels=c("70/30",
                               "90/10"),
                      values=c("#0077b6", "#03045e"), 
                      guide = guide_legend(reverse=TRUE)) +
  scale_color_manual(name="Condition",
                      labels=c("70/30",
                               "90/10"),
                      values=c("#0077b6", "#03045e"),
                     guide = guide_legend(reverse=TRUE)) +
  theme(strip.text = element_text(size=12))
```

# Agency task: Agency decisions 

## Plot: Sensitivity to value of control
```{r VoC plot, fig.height = 4, fig.width = 7, unit = "in"}

VoC_plot_sub_means <- all_data %>% 
    mutate(taskHalf = case_when(trial < 158 ~ "First Half of Task",
                                trial > 157 ~ "Second Half of Task")) %>%
    group_by(age_group, taskHalf, voc, subID, dataType) %>%
    summarize(meanSubAgency = mean(agency, na.rm = T))

VoC_plot_means <- VoC_plot_sub_means %>% 
    group_by(age_group, taskHalf, voc, dataType) %>%
    summarize(meanAgency = mean(meanSubAgency, na.rm = T),
              seAgency = sd(meanSubAgency / sqrt(n())))

#plot
VoC_plot <- ggplot(VoC_plot_means, aes(x = voc, y = meanAgency, color = age_group)) +
      facet_grid(cols = vars(taskHalf), rows = vars(dataType)) +
    geom_point(aes(color = age_group)) + 
    geom_errorbar(aes(color = age_group, ymin = meanAgency - seAgency, ymax = meanAgency + seAgency), width = .1) + 
    geom_line(aes(group = age_group)) +
    voc_theme() + 
    scale_color_manual(values=c("#702963", "#c00000", "#ffae42"), name = "Age Group") +
    xlab("Value of Choice (VoC)") +
    ylab("Proportion Agency Choices") +
    geom_hline(yintercept = .5, linetype = "dashed") +
    geom_vline(xintercept = 0, linetype = "dashed")
VoC_plot
```

## Compute average error in agency predictions
```{r posterior predictive error}

sim_means <- all_data %>%
    filter(dataType == "Model Simulation") %>%
    group_by(subID, age, age_group, trial, condition, tokenOffer) %>%
    summarize(simMeanAgency = mean(agency, na.rm = T))

VoC_sub_error <- all_data %>%
    filter(dataType == "Empirical") %>%
    select(subID, age, age_group, trial, condition, tokenOffer, agency, voc) %>%
    left_join(sim_means, by = c("subID", "age", "age_group", "trial", "condition", "tokenOffer")) %>%
    mutate(error = simMeanAgency - agency,
           absError = abs(error),
           taskHalf = case_when(trial < 158 ~ "First Half of Task",
                                trial > 157 ~ "Second Half of Task")) 

VoC_mean_error <- VoC_sub_error %>%
    group_by(age_group, taskHalf, voc) %>%
    summarize(meanError = mean(error, na.rm = T),
              seError= sd(error / sqrt(n())),
              meanAbsError = mean(absError, na.tm = T)
              )
```

## Plot: Agency decision model error
```{r VoC error plot, fig.height = 4, fig.width = 7, unit = "in"}

#plot
VoC_error_plot <- ggplot(VoC_mean_error, aes(x = voc, y = meanError, color = age_group)) +
    facet_grid(cols = vars(taskHalf)) +
    geom_point(aes(color = age_group)) + 
    geom_errorbar(aes(color = age_group, ymin = meanError - seError, ymax = meanError + seError), width = .1) + 
    geom_line(aes(group = age_group)) +
    voc_theme() + 
    scale_color_manual(values=c("#702963", "#c00000", "#ffae42"), name = "Age Group") +
    xlab("Value of Choice (VoC)") +
    ylab("Mean Error in Model Predictions") +
    geom_hline(yintercept = 0, linetype = "dashed") +
    geom_vline(xintercept = 0, linetype = "dashed")
VoC_error_plot
```

## Summary stats: Sensitivity to value of control
```{r VoC summary stats}

# What proportion of trials did they choose agency when VoC was 0?
VoC_zero_means_sub <- all_data %>% 
    filter(voc == 0) %>%
    group_by(age_group, subID, dataType) %>%
    summarize(meanSubAgency = mean(agency, na.rm = T))

VoC_zero_means_ageGrp <- VoC_zero_means_sub %>%
    group_by(age_group, dataType) %>%
    summarize(meanAgency = mean(meanSubAgency, na.rm = T),
              seAgency = sd(meanSubAgency / sqrt(n())))
VoC_zero_means_ageGrp


VoC_zero_means <- VoC_zero_means_sub %>%
    ungroup() %>%
    summarize(meanAgency = mean(meanSubAgency, na.rm = T),
              seAgency = sd(meanSubAgency / sqrt(n())))
VoC_zero_means

# What was average VoC when participants had 50% chance of choosing agency?
VoC_chance_means_sub <- all_data %>% 
    group_by(age_group, subID, voc, dataType) %>%
    summarize(meanSubAgency = mean(agency, na.rm = T)) 

#identify when 50% line is crossed
VoC_chance_means <- VoC_chance_means_sub %>% 
    group_by(age_group, voc, dataType) %>%
    summarize(meanAgency = mean(meanSubAgency, na.rm = T),
              seAgency = sd(meanSubAgency / sqrt(n()))) %>%
    ungroup() %>%
    group_by(age_group, dataType) %>%
    mutate(priorMeanAgency = lag(meanAgency),
           threshold = case_when(meanAgency > .5 & priorMeanAgency < .5 ~ 1)) %>%
    filter(threshold == 1)
VoC_chance_means
```
